primitives: move the Poseidon tests into a tests subdir
