+++ crates/ragu_primitives/src/lib.rs
@@ -28,8 +28,6 @@ mod invertible;
 pub mod io;
 mod point;
 pub mod poseidon;
-#[cfg(test)]
-mod poseidon_halo2_vectors;
 pub mod promotion;
 mod sendable;
 mod simulator;
